Fixed repopulation error for visible radio buttons on new subject form
This issue was cuased by line the line 11 in create_subject.php which has no been modified. This assumed the POST value was already set, and when we enfoced the int type on the new $visbile value, it returned 0 for false becuase the visible value was not posted and thus we ended up pushing this value to the session as if it was posted as a 0 to begin with.

The fix was to simply only set $visible if the post value for visible was also set and posted.