Progress: added page and form for editing subjects
New styles were added. Classes for bold, inline, block and normal.

New Subject form now has full indication of what values were entered
incorectly and need to be fixed.

Manage content has been refactored and now displays more information
about the current subject selected.

Edit subject was created. It automatically populates the current form.
Submit is currently not implmented correctly.

When user clicks on edit, the id fo the current subject is sent via GET
to the edit_subject.php, which than includes find_current_subject.php
which detects GET values, and if it finds any it automatically returns a
assoc array of data from the database for the row of the same id in the
GET. Values to work with for subject will be $current_subject.