Editing a subject now shows correct messages upon submission
When submiting identical values for change in (edit_subject.php), error message is no longer an error, but a notice that values were the same and thus no changes were made.

This is becuase if a query is sucessful, doesn't mean anything actually happened. This is determined instead by the number of rows affected.

Minor visual refactor in visibility radio buttonbs in (edit_subject.php).