Working on delete subject functionality
Code refactor in (manage_content.php). Code for subject overview is no longer concatenated in string $output and than echoed, instead the inside of the if condition is closed with ending php tags and raw but conditional html is output if condition is true.

NEXT STEPS working on custon validation, so that if user inputs the correct menu name, than they are able to delete the subject, as a security measure.