Subject deletion added
Refactor, delete modal now has no form, only a anchor tag that links to (delete_subject.php) with get paramater of page or subject = id.

This anchor tag is only available to click if input value to text field matches the menu name of the subject you are trying to delete.

All pages of the subject you are trying to delete are also deleted.