Fixed responsive position adjustment for moving pages between subject
This is done by calling adjust_position_for_page_addition in the subject for which the page is moving into. And respectively calling adjust_position_for_page_deletion in the subject for which the page is moving out of.

Also added notification message in nav if no subjects exist to be displayed.