Added htaccess and fixed bug issue when moving pages between subjects
Htacess rule now hides .php extension form files in url. This however does not yet apply to the extension in regards to query strings.

A bug existed. When moving pages between subjects, the subject from which the page was moved would reorder its pages incorectly. This was resolved by passing in both the new position and the old position for the page being moved to make reorder logic more accurate.