Refactor sitewide header and displaying login link on homepage
Removed function for generating header and moved header to a seperate included file called (header.php) in included layouts.

Changed all references of generate_header() function to includes for new header file.

Navigation conditions (for example, weather to output and display hamburger icon for mobile nav on pages that have navigation for content) now reply on ($context) assoc array being defined at the start of the script.

Note: not having this array defined does not throw error, as (header.php) checks to see if it is set first.

Index page now displays a desktop and mobile nav link to the login page.

Added style (side-padding-buffer) to increase side paddings to 30px. This is used on login link's anchor tag on homepage for design.